feat : 첫 input focus
