+++ src/component/input/CardLabeledInput.tsx
@@ -2,7 +2,8 @@ import FormField from '../FormField';
 import CardInput from './CardInput';
 import type { CardInputProps } from '../../types/CardInputTypes';
 import type { CardInputConfig } from '../../types/CardConfigTypes';
-import { useRef, useEffect } from 'react';
+import { useEffect } from 'react';
+import useFocus from '../../hook/useFocus';
 
 interface ValueProps {
   [key: string]: string;
@@ -21,18 +22,11 @@ interface CardInputGroupProps {
 }
 
 const CardLabeledInput = ({ config, errorMessage, handleCardInput, isErrors, value }: CardInputGroupProps) => {
-  const inputRefsArray = config.inputKeys.map((key) => [key, useRef<HTMLInputElement>(null)]);
-  const inputRefsObject = Object.fromEntries(inputRefsArray);
+  const { inputRefsObject, handleFocus } = useFocus(config.inputKeys);
 
-  const handleFocus = (key: keyof CardInputProps, maxLength: number) => {
-    const currentIndex = config.inputKeys.indexOf(key);
-    const nextKey = config.inputKeys[currentIndex + 1];
-
-    if (!nextKey) return;
-    if (inputRefsObject[key].current.value.length === maxLength) {
-      inputRefsObject[nextKey].current?.focus();
-    }
-  };
+  useEffect(() => {
+    inputRefsObject[config.inputKeys[0]].current?.focus();
+  }, []);
 
   return (
     <FormField label={config.label} errorMessage={errorMessage} id={config.id}>
